fix: circular dependency warning by moving STACK_SCREENS from Welcome.js into Confirmation.js
